refactor(p1): extract ChatInputAttachments.kt — chips, pickers, editor, cropper
Second ChatInput god-class extraction (~500 lines).

What moves to ChatInputAttachments.kt (542 lines new):
- internal @Composable MediaFileInputRow — the per-attachment chip row
- internal @Composable FilesPicker — the row of 5 pick buttons
- internal @Composable FullScreenEditor — fullscreen text editor dialog
- internal @Composable useCropLauncher — UCrop ActivityResult hook
- internal @Composable ImagePickButton — gallery pick button
- public @Composable TakePicButton / VideoPickButton / AudioPickButton /
  FilePickButton (visibility preserved — they were top-level `public fun`
  before, even though they have no external callers; demoting is out of
  scope for this refactor)
- private helpers: ImageAttachmentPreview, ImageAttachmentStatus.dotColor,
  AttachmentChip, AttachmentLeadingIcon, attachmentNameFromUrl, BigIconTextButton,
  BigIconTextButtonPreview

ChatInput.kt: 3081 → 2578 lines (-16%, -503 lines).

Side cleanup — 27 dead imports dropped from ChatInput.kt (verified zero
non-import refs in the post-strip file):
  ActivityResultLauncher, LocalIndication, horizontalScroll,
  MutableInteractionSource, FlowRow, safeDrawingPadding, wrapContentWidth,
  BasicAlertDialog, ProvideTextStyle, Role/role/semantics, DialogProperties,
  toFile, UCrop, UCropActivity, Camera01, Files02, Image02, MusicNote01,
  Video01, ImageAttachmentStatus/Kind, getCurrentChatModel, PermissionCamera,
  PermissionManager, rememberPermissionState.

Verification: :app:compileDebugKotlin --offline → BUILD SUCCESSFUL (initial
clean was too aggressive — `withContext`/`Dispatchers`/`imePadding`/`clip`/
`CircleShape`/`fillMaxHeight` were still used at other ChatInput sites and
got restored before commit).
