+++ app/src/main/java/me/rerere/rikkahub/ui/components/ai/ChatInputAttachments.kt
@@ -1,0 +1,614 @@
+package me.rerere.rikkahub.ui.components.ai
+
+import android.app.Activity
+import android.content.Intent
+import android.graphics.Bitmap
+import android.net.Uri
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.LocalIndication
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.horizontalScroll
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.FlowRow
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.imePadding
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.safeDrawingPadding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.widthIn
+import androidx.compose.foundation.layout.wrapContentWidth
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.BasicAlertDialog
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.ProvideTextStyle
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TextField
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.produceState
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.semantics.role
+import androidx.compose.ui.semantics.semantics
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.util.fastForEach
+import androidx.compose.ui.window.DialogProperties
+import androidx.core.net.toFile
+import androidx.core.net.toUri
+import coil3.compose.AsyncImage
+import com.dokar.sonner.ToastType
+import com.yalantis.ucrop.UCrop
+import com.yalantis.ucrop.UCropActivity
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+import me.rerere.ai.provider.ProviderSetting
+import me.rerere.ai.ui.UIMessagePart
+import me.rerere.common.android.appTempFolder
+import me.rerere.hugeicons.HugeIcons
+import me.rerere.hugeicons.stroke.Camera01
+import me.rerere.hugeicons.stroke.Cancel01
+import me.rerere.hugeicons.stroke.Files02
+import me.rerere.hugeicons.stroke.Image02
+import me.rerere.hugeicons.stroke.MusicNote01
+import me.rerere.hugeicons.stroke.Video01
+import me.rerere.rikkahub.R
+import me.rerere.rikkahub.data.ai.vision.ImageAttachmentStatus
+import me.rerere.rikkahub.data.ai.vision.ImageAttachmentStatusKind
+import me.rerere.rikkahub.data.ai.vision.ImageAttachmentValidator
+import me.rerere.rikkahub.data.datastore.findProvider
+import me.rerere.rikkahub.data.datastore.getCurrentChatModel
+import me.rerere.rikkahub.data.files.FilesManager
+import me.rerere.rikkahub.ui.components.ui.permission.PermissionCamera
+import me.rerere.rikkahub.ui.components.ui.permission.PermissionManager
+import me.rerere.rikkahub.ui.components.ui.permission.rememberPermissionState
+import me.rerere.rikkahub.ui.components.ui.workspaceColors
+import me.rerere.rikkahub.ui.context.LocalSettings
+import me.rerere.rikkahub.ui.context.LocalToaster
+import me.rerere.rikkahub.ui.hooks.ChatInputState
+import org.koin.compose.koinInject
+import java.io.File
+
+/**
+ * Composer attachment chips, file/image/video/audio picker buttons, full-screen
+ * text editor, and the UCrop launcher hook. Extracted from the ChatInput god
+ * class so the main composable stays focused on layout + state.
+ *
+ * Only `MediaFileInputRow`, `FilesPicker`, `FullScreenEditor`, `ImagePickButton`,
+ * and `useCropLauncher` are called from the main `ChatInput()` composable —
+ * those are `internal`. Other helpers stay `private` / preserve their original
+ * top-level public visibility (TakePicButton / VideoPickButton / AudioPickButton
+ * / FilePickButton were `public fun` before the extraction and stay that way).
+ */
+
+@Composable
+internal fun MediaFileInputRow(
+    state: ChatInputState,
+) {
+    val filesManager: FilesManager = koinInject()
+    val settings = LocalSettings.current
+    val toaster = LocalToaster.current
+    val managedFiles by filesManager.observe().collectAsState(initial = emptyList())
+    val displayNameByRelativePath = remember(managedFiles) {
+        managedFiles.associate { it.relativePath to it.displayName }
+    }
+    val displayNameByFileName = remember(managedFiles) {
+        managedFiles.associate { it.relativePath.substringAfterLast('/') to it.displayName }
+    }
+
+    fun removePart(part: UIMessagePart, url: String) {
+        state.messageContent = state.messageContent.filterNot { it == part }
+        if (state.shouldDeleteFileOnRemove(part)) {
+            filesManager.deleteChatFiles(listOf(url.toUri()))
+        }
+    }
+
+    Row(
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(horizontal = 6.dp, vertical = 6.dp)
+            .horizontalScroll(rememberScrollState())
+    ) {
+        state.messageContent.fastForEach { part ->
+            when (part) {
+                is UIMessagePart.Image -> {
+                    val status by produceState(
+                        ImageAttachmentValidator.checking(),
+                        part.url,
+                        settings.chatModelId,
+                        settings.ocrModelId,
+                        settings.providers,
+                    ) {
+                        value = withContext(Dispatchers.IO) {
+                            ImageAttachmentValidator.inspectImage(part, settings)
+                        }
+                    }
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = "image",
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = {
+                            ImageAttachmentPreview(
+                                url = part.url,
+                                status = status,
+                                onStatusClick = {
+                                    if (status.blocksSend) {
+                                        toaster.show(status.message, type = ToastType.Error)
+                                    }
+                                }
+                            )
+                        },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                is UIMessagePart.Video -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = "video",
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = { AttachmentLeadingIcon(icon = HugeIcons.Video01) },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                is UIMessagePart.Audio -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = "audio",
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = { AttachmentLeadingIcon(icon = HugeIcons.MusicNote01) },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                is UIMessagePart.Document -> {
+                    AttachmentChip(
+                        title = attachmentNameFromUrl(
+                            url = part.url,
+                            fallback = part.fileName,
+                            displayNameByRelativePath = displayNameByRelativePath,
+                            displayNameByFileName = displayNameByFileName
+                        ),
+                        leading = { AttachmentLeadingIcon(icon = HugeIcons.Files02) },
+                        onRemove = { removePart(part, part.url) }
+                    )
+                }
+
+                else -> Unit
+            }
+        }
+    }
+}
+
+@Composable
+private fun ImageAttachmentPreview(
+    url: String,
+    status: ImageAttachmentStatus,
+    onStatusClick: () -> Unit,
+) {
+    Box(
+        modifier = Modifier
+            .size(34.dp)
+            .clickable(
+                enabled = status.blocksSend,
+                onClick = onStatusClick,
+            )
+    ) {
+        Surface(
+            modifier = Modifier.fillMaxSize(),
+            shape = RoundedCornerShape(10.dp),
+            color = MaterialTheme.colorScheme.surfaceContainerHigh,
+        ) {
+            AsyncImage(
+                model = url,
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier.fillMaxSize()
+            )
+        }
+        Surface(
+            modifier = Modifier
+                .size(8.dp)
+                .align(Alignment.TopEnd),
+            shape = CircleShape,
+            color = status.dotColor(),
+            tonalElevation = 0.dp,
+            shadowElevation = 0.dp,
+        ) {}
+    }
+}
+
+@Composable
+private fun ImageAttachmentStatus.dotColor(): Color = when (kind) {
+    ImageAttachmentStatusKind.CHECKING -> workspaceColors().muted
+    ImageAttachmentStatusKind.READY -> Color(0xFF2EAD5B)
+    ImageAttachmentStatusKind.FALLBACK -> Color(0xFFFFB020)
+    ImageAttachmentStatusKind.BLOCKED -> MaterialTheme.colorScheme.error
+}
+
+@Composable
+private fun AttachmentChip(
+    title: String,
+    leading: @Composable () -> Unit,
+    onRemove: () -> Unit,
+) {
+    val workspace = workspaceColors()
+    Surface(
+        shape = RoundedCornerShape(8.dp),
+        tonalElevation = 0.dp,
+        shadowElevation = 0.dp,
+        color = workspace.paper,
+        contentColor = workspace.ink,
+        border = BorderStroke(1.dp, workspace.hairline)
+    ) {
+        Row(
+            modifier = Modifier
+                .height(44.dp)
+                .padding(start = 8.dp, end = 6.dp),
+            verticalAlignment = Alignment.CenterVertically,
+            horizontalArrangement = Arrangement.spacedBy(8.dp)
+        ) {
+            leading()
+            Text(
+                text = title,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+                style = MaterialTheme.typography.bodyMedium,
+                modifier = Modifier.widthIn(min = 40.dp, max = 180.dp),
+            )
+            Box(
+                modifier = Modifier
+                    .clip(CircleShape)
+                    .size(26.dp)
+                    .clickable(onClick = onRemove),
+                contentAlignment = Alignment.Center
+            ) {
+                Icon(
+                    imageVector = HugeIcons.Cancel01,
+                    contentDescription = null,
+                    tint = workspace.muted,
+                    modifier = Modifier.size(16.dp)
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun AttachmentLeadingIcon(
+    icon: ImageVector,
+) {
+    val workspace = workspaceColors()
+    Surface(
+        modifier = Modifier.size(32.dp),
+        shape = RoundedCornerShape(6.dp),
+        color = workspace.row,
+        contentColor = workspace.muted,
+        border = BorderStroke(1.dp, workspace.hairline),
+    ) {
+        Box(
+            modifier = Modifier.fillMaxSize(),
+            contentAlignment = Alignment.Center
+        ) {
+            Icon(
+                imageVector = icon,
+                contentDescription = null,
+                tint = workspace.muted
+            )
+        }
+    }
+}
+
+private fun attachmentNameFromUrl(
+    url: String,
+    fallback: String,
+    displayNameByRelativePath: Map<String, String>,
+    displayNameByFileName: Map<String, String>,
+): String {
+    val parsed = runCatching { url.toUri() }.getOrNull()
+    val relativePath = parsed?.path?.substringAfter("/files/", missingDelimiterValue = "")?.takeIf { it.isNotBlank() }
+    if (relativePath != null) {
+        displayNameByRelativePath[relativePath]?.let { return it }
+    }
+
+    val storedFileName = parsed?.lastPathSegment?.substringAfterLast('/')?.takeIf { it.isNotBlank() }
+    if (storedFileName != null) {
+        displayNameByFileName[storedFileName]?.let { return it }
+        return storedFileName
+    }
+
+    return fallback
+}
+
+@Composable
+internal fun FilesPicker(
+    onTakePic: () -> Unit,
+    onPickImage: () -> Unit,
+    onPickVideo: () -> Unit,
+    onPickAudio: () -> Unit,
+    onPickFile: () -> Unit,
+) {
+    val settings = LocalSettings.current
+    val provider = settings.getCurrentChatModel()?.findProvider(providers = settings.providers)
+
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp), verticalArrangement = Arrangement.spacedBy(16.dp)
+    ) {
+        FlowRow(
+            modifier = Modifier.fillMaxWidth(),
+            horizontalArrangement = Arrangement.spacedBy(16.dp),
+            verticalArrangement = Arrangement.spacedBy(4.dp),
+        ) {
+            TakePicButton(onLaunchCamera = onTakePic)
+
+            ImagePickButton(onClick = onPickImage)
+
+            if (provider != null && provider is ProviderSetting.Google) {
+                VideoPickButton(onClick = onPickVideo)
+
+                AudioPickButton(onClick = onPickAudio)
+            }
+
+            FilePickButton(onClick = onPickFile)
+        }
+    }
+}
+
+@Composable
+internal fun FullScreenEditor(
+    state: ChatInputState, onDone: () -> Unit
+) {
+    BasicAlertDialog(
+        onDismissRequest = {
+            onDone()
+        },
+        properties = DialogProperties(
+            usePlatformDefaultWidth = false, decorFitsSystemWindows = false
+        ),
+    ) {
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .safeDrawingPadding()
+                .imePadding(),
+            verticalArrangement = Arrangement.Bottom
+        ) {
+            Surface(
+                modifier = Modifier
+                    .widthIn(max = 800.dp)
+                    .fillMaxHeight(0.9f),
+                shape = RoundedCornerShape(topStart = 16.dp, topEnd = 16.dp)
+            ) {
+                Column(
+                    modifier = Modifier
+                        .padding(8.dp)
+                        .fillMaxSize(),
+                    horizontalAlignment = Alignment.End,
+                    verticalArrangement = Arrangement.spacedBy(8.dp),
+                ) {
+                    Row {
+                        TextButton(
+                            onClick = {
+                                onDone()
+                            }) {
+                            Text(stringResource(R.string.chat_page_save))
+                        }
+                    }
+                    TextField(
+                        state = state.textContent,
+                        modifier = Modifier
+                            .padding(bottom = 2.dp)
+                            .fillMaxSize(),
+                        shape = RoundedCornerShape(32.dp),
+                        placeholder = {
+                            Text(stringResource(R.string.chat_input_placeholder))
+                        },
+                        colors = TextFieldDefaults.colors().copy(
+                            unfocusedIndicatorColor = Color.Transparent,
+                            focusedIndicatorColor = Color.Transparent,
+                            focusedContainerColor = Color.Transparent,
+                            unfocusedContainerColor = Color.Transparent,
+                        ),
+                    )
+                }
+            }
+        }
+    }
+}
+
+@Composable
+internal fun useCropLauncher(
+    onCroppedImageReady: (Uri) -> Unit, onCleanup: (() -> Unit)? = null
+): Pair<ActivityResultLauncher<Intent>, (Uri) -> Unit> {
+    val context = LocalContext.current
+    var cropOutputUri by remember { mutableStateOf<Uri?>(null) }
+
+    val cropActivityLauncher = rememberLauncherForActivityResult(
+        ActivityResultContracts.StartActivityForResult()
+    ) { result ->
+        if (result.resultCode == Activity.RESULT_OK) {
+            cropOutputUri?.let { croppedUri ->
+                onCroppedImageReady(croppedUri)
+            }
+        }
+        // Clean up crop output file
+        cropOutputUri?.toFile()?.delete()
+        cropOutputUri = null
+        onCleanup?.invoke()
+    }
+
+    val launchCrop: (Uri) -> Unit = { sourceUri ->
+        val outputFile = File(context.appTempFolder, "crop_output_${System.currentTimeMillis()}.jpg")
+        cropOutputUri = Uri.fromFile(outputFile)
+
+        val cropIntent = UCrop.of(sourceUri, cropOutputUri!!).withOptions(UCrop.Options().apply {
+            setFreeStyleCropEnabled(true)
+            setAllowedGestures(
+                UCropActivity.SCALE, UCropActivity.ROTATE, UCropActivity.NONE
+            )
+            setCompressionFormat(Bitmap.CompressFormat.PNG)
+        }).withMaxResultSize(4096, 4096).getIntent(context)
+
+        cropActivityLauncher.launch(cropIntent)
+    }
+
+    return Pair(cropActivityLauncher, launchCrop)
+}
+
+@Composable
+internal fun ImagePickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.Image02, null)
+    }, text = {
+        Text(stringResource(R.string.photo))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+fun TakePicButton(onLaunchCamera: () -> Unit = {}) {
+    val cameraPermission = rememberPermissionState(PermissionCamera)
+
+    // 使用权限管理器包装
+    PermissionManager(
+        permissionState = cameraPermission
+    ) {
+        BigIconTextButton(icon = {
+            Icon(HugeIcons.Camera01, null)
+        }, text = {
+            Text(stringResource(R.string.take_picture))
+        }) {
+            if (cameraPermission.allRequiredPermissionsGranted) {
+                onLaunchCamera()
+            } else {
+                // 请求权限
+                cameraPermission.requestPermissions()
+            }
+        }
+    }
+}
+
+@Composable
+fun VideoPickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.Video01, null)
+    }, text = {
+        Text(stringResource(R.string.video))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+fun AudioPickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.MusicNote01, null)
+    }, text = {
+        Text(stringResource(R.string.audio))
+    }) {
+        onClick()
+    }
+}
+
+@Composable
+fun FilePickButton(onClick: () -> Unit = {}) {
+    BigIconTextButton(icon = {
+        Icon(HugeIcons.Files02, null)
+    }, text = {
+        Text(stringResource(R.string.upload_file))
+    }) {
+        onClick()
+    }
+}
+
+
+@Composable
+private fun BigIconTextButton(
+    modifier: Modifier = Modifier,
+    icon: @Composable () -> Unit,
+    text: @Composable () -> Unit,
+    onClick: () -> Unit,
+) {
+    val interactionSource = remember { MutableInteractionSource() }
+    Column(
+        modifier = modifier
+            .clip(RoundedCornerShape(8.dp))
+            .clickable(
+                interactionSource = interactionSource, indication = LocalIndication.current, onClick = onClick
+            )
+            .semantics {
+                role = Role.Button
+            }
+            .wrapContentWidth(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.spacedBy(2.dp)) {
+        Surface(
+            tonalElevation = 2.dp, shape = RoundedCornerShape(8.dp)
+        ) {
+            Box(
+                modifier = Modifier.padding(horizontal = 32.dp, vertical = 16.dp)
+            ) {
+                icon()
+            }
+        }
+        ProvideTextStyle(MaterialTheme.typography.bodySmall) {
+            text()
+        }
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun BigIconTextButtonPreview() {
+    Row(
+        modifier = Modifier.padding(16.dp)
+    ) {
+        BigIconTextButton(icon = {
+            Icon(HugeIcons.Image02, null)
+        }, text = {
+            Text(stringResource(R.string.photo))
+        }) {}
+    }
+}
